+++ src/vcdiff_read.c
@@ -10,7 +10,7 @@ vcdiff_read_rc_t vcdiff_read_byte (uint8_t *dst, const uint8_t **input, size_t *
 	return VCDIFF_READ_DONE;
 }
 
-vcdiff_read_rc_t vcdiff_read_int (uint32_t *dst, const uint8_t **input, size_t *input_remainder) {
+vcdiff_read_rc_t vcdiff_read_int (size_t *dst, const uint8_t **input, size_t *input_remainder) {
 	vcdiff_read_rc_t rc = VCDIFF_READ_CONT;
 
 	while (rc == VCDIFF_READ_CONT && *input_remainder > 0) {
